cleanup_archive/requirements.agentcore.txt
# AgentCore MCP Server Dependencies (Simplified)
# Core MCP protocol library (official AWS AgentCore compatible)
mcp>=1.0.0

# FastAPI for HTTP server (if needed)
fastapi>=0.104.0
uvicorn>=0.24.0

# AWS dependencies for cost tracking and monitoring
boto3>=1.39.7
botocore>=1.39.7

# Async support
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0

# Data handling
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Logging and monitoring
structlog>=23.2.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0

# Manufacturing tool dependencies (simplified for AgentCore)
openai>=1.0.0
anthropic>=0.7.0

# Strands framework for manufacturing agent
strands-agents>=1.5.0
strands-agents-tools>=0.2.4

# Basic optimization solvers (no compilation required)
numpy>=1.24.0
scipy>=1.11.0

# Async support
nest-asyncio>=1.5.0
